backend/requirements.txt
# FastAPI and dependencies
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Email
aiosmtplib==3.0.1
jinja2==3.1.3

# File handling
python-magic==0.4.27
aiofiles==23.2.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
pytest-cov==4.1.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3 